hoan thanh chuc nang ngon ngu

+++ ThuongMaiDienTu/database/migrations/2026_05_27_000002_create_attribute_translations_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('attribute_translations', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('attribute_id');
+            $table->unsignedInteger('attribute_id');
             $table->string('locale', 10);
             $table->string('name');
             $table->text('description')->nullable();
